I05-29 Clean up and add logging to arpes.py
To help investigate the cause of the "configure only" bug additional
logging is needed here. Also this class contained some progress
reporting relating to the command queue which is no longer used on i05
so it is removed. There are also 3 other methods which seem to have
nothing to do with running ARPES scans which are also removed.

Change-Id: I1cbeb92131b277d768af4a22e2bac9b793f2d4e9