gdalogpanel Keep more events and only restrict size in the viewer
It seems the main restriction on number of events is the number shown in
the GUI. If we keep many more events in memory and restrict the number
at the time the GUI is populated, we avoid the situation where no
messages are visible in the panel because it's entirely full of
lower-level events which are hidden.

It might be that we could now remove the "expire by priority" option
altogether but I'll leave it as it is for now.

With increased numbers of events, performance results are still good,
and it looks like now the number of events is more or less irrelevant:

Testing with min log level WARN
200000 logging messages sent;   10 UI events processed in  155
milliseconds;  0.001 milliseconds per message

Testing with min log level INFO
200000 logging messages sent;   21 UI events processed in  577
milliseconds;  0.003 milliseconds per message

Testing with min log level ALL
200000 logging messages sent;   25 UI events processed in  720
milliseconds;  0.004 milliseconds per message
200000 logging messages sent;   21 UI events processed in  803
milliseconds;  0.004 milliseconds per message
200000 logging messages sent;   20 UI events processed in  759
milliseconds;  0.004 milliseconds per message

Change-Id: I1d1c878ab3d61a7d17545dbe84713a101c115997