Remove PackageMaker and dependency on org.apache.bcel it required
As this class is a stand alone Java application with no references in
the rest of the codebase it should be safe to remove.

It appears to allow a class file to be copied to a new location by
substituting the package name separators "." with directory separators
"\" to maintain the required java directory structure in the new
location.

Change-Id: I459abff8ace15d0d0f8ef149b4bf38b356bcfb72