Make MappingScanSubmitter use the new axis manager to set axis names
This had the side effect of changing the name stored in the scan path
model and hence in the points generated by the PathInfoCalculatorJob.
These have been changed to cast IPositions to Points, allowing the getX()
and getY() methods to be used instead of getValue("x") which relies on
the axis name.

Change-Id: I01efd2e72bc6a8249456b480a1c17065157fdee5