DAQ-196 Use less aggressive server killing logic
In order to protect against accidents (e.g. inadvertently killing the 
live beamline server) the start/restart/stop logic has been modified 
as follows:

When entering gda --start servers you will only be allowed to proceed if
there are no existing Object, Name, Channel or Log servers running on 
the target machine; i.e. 'start' means start from scratch.

When entering gda --restart/--stop servers, the command will only 
succeed if all the servers of any type that would be terminated:
    1. were previously started by the user issuing the command (in 
       effect gda2 on control machines).
    2. were previously started from the same main server target you are 
       trying to restart/stop; i.e. if v9.2.1 is already running you 
       won't be allowed to restart from a deployment that points to 
       v9.2.5 without manually stopping/clearing up the other server 
       first (N.B. successive dev server exports under the same /servers 
       directory will be considered 'the same' for these purposes).
    3. were previously started from the same deployment (i.e. where the 
       gda link points to) as the one you are trying to restart/stop.

This should prevent inadvertent server termination in the same way that 
the gda python script used to.

To facilitate this, any resulting errors are also written to the server 
startup file so that they can be retrieved and displayed in the terminal 
via a separate modification to 'lookforfile'

Change-Id: I526b169d51c362eb7a833dae8593dffa8399559a