I05-29 Clean up and add logging to arpes.py
To help investigate the cause of the "configure only" bug additional
logging is needed here. Also this class contained some progress
reporting relating to the command queue which is no longer used on i05
so it is removed. There are also 3 other methods which seem to have
nothing to do with running ARPES scans which are also removed.

This is the equivalent of commit 8ff2b44 on master but as the script has
moved since then this is here to add the fix to 8.master

Change-Id: I27432896958a15cd2a8424934966bee3253b6aac