DAQ-241: Fix split package preventing Jython startup (once import * works)
In GDA9 there is no classpath and (because of OSGi) protected methods 
may only override methods of a superclass if it is in BOTH the same 
package and the same plugin as the subclass, or a Verify Error is 
generated. Once the 'from XX import * ' syntax is working, this happens 
on Jython initialisation for the Polarimeter classes as they are in 
uk.ac.gda.epics whereas their superclass is in uk.ac.gda.core. So move 
these classes to uk.ac.gda.core and extract an interface to 
uk.ac.gda.api for the moved classes to use to refer to members that 
remain in epics.

Change-Id: I53f7b862fc247755bca531cdc16de337f94fb964