DAQ-241: Fix split package preventing Jython startup (once import * works))
In GDA9 there is no classpath and (because of OSGi) protected methods 
may only override methods of a superclass if it is in BOTH the same 
package and the same plugin as the subclass, or a Verify Error is 
generated. Once the 'from XX import * ' syntax is working, this happens 
on Jython initialisation for the Polarimeter classes as they are in 
uk.ac.gda.epics whereas their superclass is in uk.ac.gda.core. So move 
these classes to uk.ac.gda.core and implement a newly extracted 
interface on the remaining epics class that they still refer to.

Change-Id: I10e56dbbd6bc0f36c609bbb0484d21f80113ef54