I09-70 Add buttons to run the sequence file or copy the command to the clipboard
This adds the buttons to run the sequence now or copy the command to run
it to the clipboard for inspection or use in scripts. It enables extra
detectors to be configured using the extraDetectors object in the
Jython namespace, which is just a string appended to the end of the
analyserscan command.

Some of the code for this was already implemented so was re-enabled

Change-Id: I8bd39d04c7dc791e2d95c87c42ed9f07e06a496f