Fix enable/disable of autoscaling for live plots, and KE/BE switching
This fix enables rescaling of the plots to be toggled on and off using
the "Rescale axis when plotted data changes" button and for this to
actually happen in the plots. There are some tricky code paths due to
the different ways the methods can be called either in response to EPICS
updates or by user interaction. The change also includes a few minor
UI changes.

The main fixes are:
- Synchronize the methods to do with updating the plots to stop user and
  EPICS calls interfering.
- Change uses of Dataset to IDataset (just tidier)
- Create a new reverseXAxis method which can ensure BE is plotted with
  x axis reversed
- Only set scales if auto-rescaling is enabled, to allow live plots to
  be zoomed
- Change from calling createPlot... to updatePlot... to stop the plotted
  data being completely replaced.

Change-Id: I36a3fa8e458a98ac5a9c0b930089ad2a7ce01306