I05-106 Add scan filename to status bar
This change means while a scan is running the current filename is
displayed in the status bar and when a scan is finished or aborted the
last scan filename is displayed instead of just the number. It handles
the case during the first point where no file is yet open by not
displaying anything.

Also:
- Organised imports

Change-Id: If3ff3674bf005cd4d43716cf7a2fd6d203056fc3