[DAQ-268] rename old ScannableNexusWrapperTest, added new test
Renamed the ScannableNexusWrapperTest to ScannableNexusWrapperScanTest
as it is not a unit test, but a test that runs a scan with lots of
ScannableNexusWrappers wrapping GDA8 Scannables.
Added new ScannableNexusWrapperTest containing unit tests for the 
methods of ScannableNexusWrapper.
Incorporates redoing Gary's abandoned change (Gerrit #4478, minor refactor
of ScannableNexusWrapperScanTest to only compound generators once).


Change-Id: I97fa40629a28d37c00667572faa5b65b520d2ac7