I05-105 Support for analyser KE ranges to be configured in Spring
This add a new class which can hold the analyser energy range data and
offer convenient methods to access it. It also provides a Python script
which can take the data provided by the vendor SES software and convert
it into a Spring XML file to instantiate this class.

The new class is immutable and serializable to allow it to be passed
over RMI to the client where it will mostly be used.

Change-Id: I85e4c5baea9b558b8983151d954dfe6b583caa94