I05-105 Add energy range support to Analyser classes
This adds a new VGScientaAnalyserEnergyRange field to the
VGScientaAnalyser and VGScientaAnalyserCamOnly classes. They now also
implement a new interface IVGScientaAnalyserRMI which will allow this
energy range object to be retrieved over RMI for the GUI to use in
validating KEs.

Also adds the appropriate Spring configuration to instantiate the
objects. The new analyser_energy_range.xml file was auto-generated using
VGScientaEnergyRangeParser.py and the RMI export and imports to Spring

Change-Id: I715f2c3c78df68ebcf294e1af3bdda4a8e605c33