Move i18 microfocus scripts into a bundle that is part of the product
Currently the scripts in uk.ac.gda.client.microfocus are copied to the 
dls-root folder in the exported product to make them available to Jython. 
In future however, they will need to be in the plugins folder (and thus 
belong to a bundle that is part one of the features that make up the 
product) to be accessed by the script location extension point system. 
Also, one of the scripts actually references the ScanPositionsTwoWay 
class which is not currently part of the product (for the same reason) 
so that script won't work anyway. 

To get round this, the scripts and the ScanPositionsTwoWay class are 
moved to the uk.ac.diamond.daq.microfocus.scan bundle which is part of 
the product and the i18 config is updated accordingly

Change-Id: I9b2a30261ae6dee65e5ea8728c035ae5d0963b02