I18-86 I18-68 : Provide table 3 data when they look more appropriate than table 1 data
This fixes loading of stage 3 map Nexus files into the Microfocus
perspective which I18 scientists use to select a map point for further
investigation with EXAFS.

This change relies on the Nexus file changes introduced by commit
97dbb26 in gda-dls-beamlines-xas.git.

This commit has not been deployed to the beamline, but a similar one has
been tested and deployed. The deployed commit is pointed to by a tag
"deployed-20160921" in the gda-xas-core repository in /dls_sw/i18.
The present commit should be somewhat more robust than the deployed
commit when loading Nexus files with certain combinations of entries.

Change-Id: I5016b95d691515caf803172956f93f35e2c24b7d