Remove the need to copy script folders to dls_root on server export
Python script folders (e.g. uk.ac.gda.core/scripts) are currently copied
to the dls_root pseudo workspace parent during server export so that
they are consistently accessible withe the case when running from
eclipse. It would be better if they remained under their associated
plugins in the exported product and the intermediate path differences
were managed by the application. This would increase the ability to
remove dls_root completely which is one of the goals in replacing the
current latching mechanism. To do this we must:

1. Include script directories in the plugin's build file so that they
get copied automatically as part of the export process.
2. Ensure that these plugins are not jarred up as part of the process.
3. Adjust the Jython configuration code so that it store the new folder
location rather than the old dls_root one.
4 Provide an alternative development export target which when used
replaces the exported scripts folder with a link back to the
corresponding folder within workspace_git so that people can still edit
the scripts in the place they expect to find them.

This part of the change updates the Jython configuration code, ensures
that the core scripts directory is in the build and modifies the
exiting extension definitions slightly.

It also uses a better method to derive the location of the Jython jar
which is independent of the type of run target allowing the link to this
to be removed also.



Change-Id: I791534f8e2c989129bfee3fdada1dd3c4fe3a020