Remove the need to copy script folders to dls_root on server export
Python script folders (e.g. uk.ac.gda.core/scripts) are currently copied
to the dls_root pseudo workspace parent during server export so that
they are consistently accessible in the case when running from
eclipse. It would be better if they remained under their associated
plugins in the exported product and the intermediate path differences
were managed by the application. This would increase the ability to
remove dls_root completely which is one of the goals in replacing the
current latching mechanism.

This part of the change 

1. Updates buckminster file to remove no longer needed copying of
folders to dls_root.
2. Adds the new target to produce the development export.
3. Adds a script to perform the script folder linking when the
development target is used.
4. Ensures that the already identified script folders don't get jarred
during export

Change-Id: Ib05ecc9ff3c4e6f1382dd65b941b6514c6da5e56