I15_1-206 (9.2pre) Update LiveMgpegPlot to use new DynamicData API.
This fixes the problem of the live Mjpeg plot not seeing any updates
after the initial frame since the January refactor.

Also, it will now wait for the first frame to arrive if the stream is
not running when the view is started, rather than requiring the user to
start the stream before opening the view.

Change-Id: I6eee7335207cc4f4e3cf74c17b9236f8931b97ea