Improve performance replace Vector by Set backed by CopyOnWriteArraySet
This is a minor first stage improvement to improving Corba push event
performance. This reduces the locking provided by Vector allowing the
subscriptions set to be iterated quicker.

Also minor logging improvements and change catch Throwable to Exception
and move the try-catch into the loop to, avoid a bad Subscription
stopping all events.

Change-Id: Ic7af8fe588fa1ad8fbdbeef73c86bc787c0ec6e2