Remove server target latching from startup and standardise gda script
We are moving to the deployment architecture described in
http://confluence.diamond.ac.uk/x/yIgQAw from v9.2 onwards. This change
removes the latching scripts and minimises the amount of code in
gdaservers so it can soon be retired.

Additionally it pushes standard config path settings down into a
defaults script in dls-config which may be overridden partially or
completely as required.

Shutdown handling is also improved to deal with processes unaffected by
SIGTERM.

Change-Id: Iaa8f05598db469a7265f2969b4644c2c4ed043ea