Make Subscription implement hashcode and equals
Also make EventSubscriber and Filter final

As Subscription objects are used to distribute events the intension is
that each EventSubscriber should only be register once. This change will
help that situation. It is still not absolutely guaranteed as neither
the EventSubscriber or Filter for implementers to override hashcode and
equals so this will rely on the Object implementation which should be
sufficient for our needs.

Change-Id: I7438af717dae9089ce4694aff5ba03805c8db9b4