I15_1-207 (9.2pre) Restore some of the lost collection strategy functionality.
Replacing raw AdBase calls with calls to the appropriate detector
functions, we restore the use of the configured collection strategy to
determine the behaviour of the detector.

Not all cameras are SoftwareStartStop, TriggerMode.Internal and
ImageMode.Single or have an an AcquireTime the same as the requested
collection time.

TODO: Add support for detector.atScanEnd & thus completeCollection.

At the moment, only the prepareForCollection and collectData parts of
the collection strategy are executed not completeCollection, so strategies
configured to save and restore state will save, but never restore that
state after a collection.

Change-Id: Ib827a264852c82c605a9be8375267ac207124197