DAQ-266 Initialise sys.modules to fix reset_namespace
Previously imported class definitions currently persist through a
reset_namespace which is incorrect. This happens because the sys.modules
list is not being reset when reset_namespace is called.

Add a call to sys.modules.clear() to GDAJythonInterpreter.initialise but
protect the calls to import site and Server port connection so that they
only happen at startup and not during reset_namspace. Also initialise
the script locations list to prevent script paths being added to it
multiple times.

Change-Id: I076636a479b328d046779cedf32fecda2147cf63