Refactor Jython initialisation ready for version 2.7
Much of the Jython initilisation is static with only some affecting the
PySystemState instance. Refactor its initilisation in
GDAJythonInterpreter to reflect this which makes reset_namespace
consistent with initials startup. It also paves the way for the switch
to Jython 2.7 which allows removal of the explicit import site call.

Also make scisoftPy script location name more specific.

The testConfiguringServerPassesPathsToInterpreter is temporarily
disabled as it is more a test of GDAJythonInterpreter and won't work
with its new structure. I will reinstate it when I have worked out how
to do so using PowerMock.

Change-Id: I0e35b7673136c6e39d1354196ddc458f2195cc33