Prepare repository bundles for JRebel
In order for a bundle to be accessible to JRebel for hot code
replacement in both local and remote mode two xml files must be added
along with some extra natures in the project file. The files are
generated automatically by the JRebel setup menus and are fixed for a
standalone application. The paths in the xml files have been
parameterised so that they are applicable to anyone's machine.

Change-Id: I9a8e8f26bb353ad51d4869789cb689d6c875b69a