I05-111 Fix race condition at the end of the scan resulting in wrong data
At the end of the scan there was a race condition between the scan
calling completeColection and the EPICS monitor update calling it. This
was needed because it used to be possible to set a number of frames less
than the current and that could cause the scan to end. I have removed
that "feature" so the monitorChanged event should never call
completeColection. Calling completeColection sets the wethinkweareincharge
flag which then changes the behaviour of other methods (especially
waitWhileBusy) resulting in this issue. I have also made the frame
counters volatile in the hope for making this slightly better. It is
still no where near thread safe and there are probably other race
conditions still. The plan is to completely remove this class and which
to using iterations in SES.

Change-Id: Ie085be0af700ddfb7e3294a142b46867614146a3