Improvements to FindableProcessorQueueTest
This test intermittently fails on Jenkins, The reason is that it is
pausing the test using a Thread.sleep while its hoping something happens
in another thread. The "solution" employed here is to reduce the delay
in TestCommand from 50ms to 5ms which is also moved here from
uk.ac.gda.test.helpers. This seems to significantly improve the
reliability.

This change also removes CommandQueueViewPluginTest which is broken and
is the only other user of TestCommand. Removing this also allows the
PluginTestHelpers class to be removed.

Also:
- Switch from deprecated junit.framework to org.junit imports
- Extract pause constant
- Make tests faster by reducing pauses from 500ms to 50ms and timeout
  from 500ms to 100ms

Change-Id: Id6b31b9a00f0c09eab6e8436b847775977b2571a