Rename uk.ac.gda.client.logging as the server needs it too
The uk.ac.gda.client.logging plugin allows the logback classloader to
see GDA classes in its dependent plugins since it acts as a fragment of
the logback jar. This ability is required on both server and client so
rename the plugin to something that isn't client specific and better
reflects its function.

Change-Id: Ie1088ae6b848e6a243ee9147ca61855ef8e0087f