DAQ-390 Re-factor existing live controls to allow the Jython button
This extracts a new interface LiveControl which provides two methods one
which creates the GUI composite for that control and one to find out the
group of that control. This will allow a new jython button to be
created. Also renames Control -> ScannablePositionerControl as there
will now be different types of controls.

Change-Id: Icc2939a44a1557c352e09ace70b636e718c64fd5