Modify EpicsPositioner to set the status flag to IDLE when the motor stops moving
The previous behaviour was:

- Set status flag to MOVING when DMOV reports that the motor starts moving
- set the flag to IDLE when INPOS reports the focus is in position

The problem was that, if the focus is moved by an arbitrary amount
(in GDA or EPICS), there is no callback from INPOS, as the motor
does not reach one of the defined positions, so a subsequent attempt 
to move the positioner is rejected because the motor is assumed to be
busy.

This change sets the flag to IDLE when DMOV reports the motor is no
longer moving - so long as no error has been reported.
The monitor on INPOS is retained.

Change-Id: Ib9fd8a7246641f66baad9a19accdeee9dfb2f883