Add ability to switch whether ZebraImpl waits when arming the detector
In EPICS support module versions 1-14 and above, Zebra behaves in a
way that is compatible with Area Detector i.e. a put with wait
to arm the detector only returns once the acquisition is complete.

In this case, GDA must call putNoWait() when arming the detector.
This change makes this behaviour configurable. 
	 
Change-Id: I02364bb75e4960b8c814071b41aac636a6c65017
(cherry picked from commit 96510577325495cf63e5b3b8581dd243d46a2860)