[DAQ-402] fix listening to IObservable in ScannableNexusWrapper
ScannableNexusWrapper implements IPositionListable to notify listeners
of the new position when the position of the wrapped scannable changes.
Unfortunately what the arguments to IObserver.update() should be is
not well defined. Previously listeners to ScannableNexusWrapper were
only notified when a ScannablePositionChangeEvent was received. But
many scannables simply send the new value as the argument. However,
some scannables send events that are not to do with position change,
e.g. a status change. So we check if the position is a value type
(i.e. a primitive wrapper or string, or array thereof).


Change-Id: Iea91daa216c2fd3ce70afc330fdb7806eb474373