MXGDA-2225 and MXGDA-2692 EpicsMCASimple: avoid race enabled by MXGDA-1720
'readingDone' is a flag that is supposed to indicate whether the read
has been completed. (Made volatile so changes are seen by other threads,
same for acquisitionDone.)

setIntFieldValue is used to send 1 to the 'read' PV, to request a read.

We then have a listener which responds to changes to the 'reading' PV,
and sets readingDone to true.

Previously we requested the read (using setIntFieldValue) and then set
readingDone to false. There was therefore a chance that we would request
the read, see a change to the 'reading' PV, set readingDone to true, and
then set readingDone to false again. This would cause GDA to hang
waiting for a read to finish when in fact it already had.

This change simply sets readingDone to false *before* requesting the read.

(Imports were re-organized automatically.)

Change-Id: I338f1186442c00055373bd75b5918673fe3484ab