DAQ-325 Replace uses of StringUtils.replace with String.replace
Also remove InitializingBean and afterPropertiesSet methods this was
used to ensure non null were set after creation, removing it will not
break any existing working code.

Remove deprecated backSlash field it was never set in any spring and the
setter was never called in code so it was always true.

This is to move towards removing spring dependencies from
uk.ac.gda.common

Change-Id: I0cc18caf986ec4d99a305e68f55d6455f7c850d6