Make ScriptProject validate path to prevent Jython error
At the moment a ScriptProject object stores an unvalidated String that
represents a path to a location of python scripts. Because of this, it
is possible for incorrect paths to be added to sys.path in Jython. This
happens when the String contains abbreviations (e.g. ../) or links
followed by abbreviations because most of the code that uses the path
only checks with File.getAbsolutePath which does not always expand these.

To prevent this add resolution of the real path before it is set so that
thereafter it is always fully correct. Also re-arrange some Jython
Interpreter code to allow the startup to continue if some of the script
locations don't resolve (with appropriate errors being logged) and 
update the corresponding tests.

Change-Id: I88b8182c9586868f9bf112359329eac7dd0714b5