B18-83 : Apply correct offset units to hardware from Stanford amplifier view.
This commit fixes a bug in the 'Stanford amplifier view, where the offset 
unit applied to hardware is different from the one selected in the gui. 
This problem was due to the offset value sent to amplifier being obtained 
from the value in the sensitivity unit combo box.
The controls to switch offset current on/off have also been replaced with 
more standard (and less confusing) radio buttons.

Change-Id: I1faa92e025b337ec0cbf30761aec005881ec849b