On startup fail write reason to looked for file
When the server is started from the startup scripts, lookforfile checks
for a specific file being written to gda_var before allowing the scripts
to complete. In the event of the scripts failing for some reason, the
cause is written to this (normally blank) file so that it can be
displayed automatically on interactive startup (dummy mode) or easily
retrieved on remote startup. This change extends this to also record
failure to start any of the object servers specified by the profile by
the application itself.

Change-Id: I42157c3ed64fb41f760b967866f9c80a6801d89b