I15_1-207 (9.3pre) Add support for Jython implemented AD runnable devices.
Add AreaDetectorRunnableDeviceProxy so we can define a bean which is
later injected with it's Jython delegate class.

Also use a new AbstractAreaDetectorRunnableDevice as base class for
other AreaDetectorRunnableDevices along with an Area Detector
implementation of a Dark Image detector.

Finally, since the new Runnable Devices need to be configured with a
name which is the name of the device, while the finder expects the name
to be unique to the bean, add a RunnableDeviceFinder class which can be
used to get at specific runnable devices whose name is the same as
another bean.

Change-Id: Ic7ac11c41329154be2b86f43cd1c0a9f0292f306