I05-104 Fix "acquire now" not working since changing to faster analyser
A new analyser implementation was introduced to improve performance see
I05-100. This caused the finder to return the wrong object when
searching by type. This fixed that issue by getting the correct analyser
by name, which is a bit less nice.

This fix won't work for i05 HR as it removes support for multiple
iterations this needs to be resolved. The best fix would be for this to
be full implemented in EPICS see BC-251. Also this script should be
removed and replaced by java see I05-108.

Change-Id: I1c3471df361f988a59cfc2d72b61f2b9fd366eae