I15_1-207 (9.3pre) Rework Jython Area Detector runnable devices.
Add AbstractAreaDetectorRunnableDeviceDelegate as a base class for all
AreaDetectorRunnableDeviceDelegate classes.

Change AreaDetectorRunnableDeviceProxy so it is the Runnable Device which
delegates its behaviour to the AreaDetectorRunnableDeviceDelegate it is
configured with. This allows the Proxy, it's Model and Info to be defined
in Spring configuration, while allowing the device specific behaviour to
be defined in Jython later and injected into the Proxy.

Improve javadoc of AbstractAreaDetectorRunnableDevice.

Change-Id: Ia5863c5cc5700518ca915b42f5ee574302d123ee