DAQ-425 Add support for binning to AreaDetectorRunnableDevice
This change gets the image size from the ArraySizeX_RBV and Y PVs
instead of from the SizeX_RBV and Y PVs. These number are after the
binning so should be correct. This still relies on one frame having
passed through the AD.

Also cache the model during configure as following the change made for
DAQ-402 it is used during createNexusObject causing a NPE.

Change-Id: Ic79f4d3b7cc5de4d75e0b0741e43859451fba07e