Remove SKIPPED_BUNDLES from GDAJythonClassloader
When we used Jython v2.5 there was a conflict between this and the v2.7
embedded in org.eclipse.scanning.command which resulted in errors. Also
the pydev bundle was in the product at the time which caused lengthy
exception traces in the Jython terminal when it tried to load certain
eclipse classes. 

To deal with this at the time, these packages were excluded from loading
by the GDAJythonClassloader. Now that we have moved to Jython 2.7, this
bodge can be removed. This stops errors with the classes recently moved
into org.eclipse.scanning.command which were inaccessible.

Change-Id: I07dfefdfafe47770b643dcd5883d1b094ec7beff