GdaRmiProxyFactoryBean: warn if client-side RMI proxy name differs from server-side object name
When FindableNameSetterPostProcessor is used, every Spring bean that implements
Findable has its setName method called with the Spring bean ID, to make sure
the object's (Findable) name matches the bean ID (this avoids you having to
explicitly define the "name" property, when you're already giving the bean an
ID).

If an RMI proxy is being created, and the proxy object implements Findable,
setName will be called on the proxy, which can result in the server-side
object's name changing - just due to a client-side proxy being created!

I did consider preventing setName method calls being forwarded to the
server-side object (by changing ClientSideIObservableMethodInterceptor). In the
end I decided against this for two reasons:

 * It's already possible to change an object's name at runtime, say from the
   Jython terminal - so disallowing this from the client seems wrong.

 * FindableNameSetterPostProcessor calls setName *and then checks this worked
   by immediately calling getName*. If setName is no longer forwarded to the
   server-side object, you can have a server-side object called "A", a
   client-side proxy called "B", setName("B") doesn't change the server-side
   object, the server-side object is still called "A", and you get a
   RuntimeException.

For now, let's just warn if the client-side proxy has a different name to the
server-side object.

Change-Id: I0e58050fccb8062646fefb00e65f9eabbd45b1db