I18-88 : Use spool dir property for MCA save location
This change relates to MCAs acquired and saved to disk through the
FluorescenceDetectorComposite class.

Previously on I18, the client would attempt to write MCAs into the root
of the visit directory (as determined by the property
"gda.data.scan.datawriter.datadir"). This worked for commissioning visit
directories, but in the case of a real visit the user does not have
write access to this location.

The user does have write access to the spool subdirectory, so instead we
dereference a property referring to this location and write the MCAs
there.

This change means that all beamlines using the
FluorenscenceDetectorComposite class should instantiate the property
"gda.fluorescenceDetector.spoolDir", typically with the value
"${gda.data}/$year$/$visit$/spool". If the property is missing, an
IllegalArgumentException ("Could not find property ...") will be thrown
when a user attempts to save an MCA to disk.

Change-Id: I233d5c07231f65b74ab4da793e21211790033c28