Fixup scanning points unit tests
The specific values put out by the RandomOffsetMutator have changed due
to a change in algorithms, and the least-significant digit of some float
values have changed due to floating-point instabilities for the
Lissajous generator.

This fixes up the hardcoded values for the Lissajous test case and
changes the means through which the RandomOffset cases are tested, by
asserting the actual values are within the maximum offset to the
expected values for the case with no random offset.

Signed-off-by: Charles Mita <charles.mita@diamond.ac.uk>