MXGDA-2534 Application Close Dialog that can notify local contacts
Server restart does not work in dummy, and there needs testing on an
actual beamline workstation.

This plugin is in gda-core to avoid cyclic dependency from
u.a.g.client.mx to u.a.g.client. Plus, from gda-core the dialog
can be extended to suit other beamlines more easily.

Change-Id: I53ccac1b6cdfff032da99d34d3a21e09d6a89d36