Don't lose the GDA_SOURCE property when a Logback LoggerContext is reset
The GDA_SOURCE property was only added to a LoggerContext once, when the GDA
process starts up.

If the LoggerContext was subsequently reset - for example, if the configuration
file(s) were changed, and were automatically reloaded - this property would be
lost.

Initially log messages would be OK:

  2017-03-21 14:09:25,006 INFO  [logserver/6925@ws341.diamond.ac.uk] gda.util.LogServer - Hi

But once the context had been reset, the process name would be missing:

  2017-03-21 14:09:52,019 INFO  [null/6925@ws341.diamond.ac.uk] gda.util.LogServer - Hi

Add a LoggerContextListener that restores the property when the context is reset.

Change-Id: Ieedfee36810117c206fe9654df8ec5b0e23f8961