[DAQ-530] Tolerate exceptions getting DeviceInformations for scannables
Modify IScannableDeviceService.getDeviceInformation() so that in the
event of a failure to get information for a device, it just logs the
failure and continues with the remaining devices.

Previously, this threw an exception which could cause a scan to fail,
even if the failing device was not needed in the scan.

Signed-off-by: Matthew Dickie <matthew.dickie@diamond.ac.uk>