Use variable for telnet port instead of hard coding
Ninety percent of beamlines use 9999 as the telnet port however, of
course, some don't. Modify the startup confirmation to use the symbol
defined in the dls-config defaults script so that those that have to be
free spirits can override its value.

Also update the example comment in p45's gda script.

Change-Id: If1d837c8e036b196ea83a7c62893d83cb866c92a