I07-66: Add server side object to push AD Array data to SDAPlotter
The TwoDArrayView does not manage regions of interest across multiple
clients, unlike the regular plot views, so this adds a way to monitor
data from the AD Array plugin using the regular plotting system that
should behave as expected.

May be redundant in GDA-9 with the Stream view, but that is
unavailable in GDA-8 and this should suit the needs of I07.

If it's truly redundant, then I'll delete it following the merge into
master.

Change-Id: Ie01f76b555db7ed5c000b4a40f8b21a6ac2454f3