Improve logging of mismatched names when using GdaRmiProxyFactoryBean
There already was a warning for this situation however I think it should
be an error as this is serious. If the names are mismatched, then events
won't work and the server side object will have it's name changed. This
is confusing to debug as then on future client restarts this issue isn't
detected as the names now match but events still won't work.

Change-Id: I1af98cd9b93b3c9f9f15fa1cf358813ee1b8a4d9