I15-1 (9.4pre) Add support for multi-value device values in a DeviceRequest.
Although the documentation suggests that we should be able to pass
Collections through the DeviceRequest.deviceValue this does not appear
to work (which is why testSerializeDeviceRequestWithDoubleList, 
testSerializeDeviceRequestWithArrayList & testSerializeDeviceRequestWithHashMap
are all @Ignored).

Instead we currently have to create a class with explicit properties, see NexusSlitsPosition, or one which contains a Collection which has the
@JsonProperty annotation added to Collection classes, see DeviceValueMultiPosition.

Ultimately we shouldn't need these classes, but for the moment this is
the most expedient way to add this required functionality.