B18-92 : Save fluorescence detector MCA data in column format
A dialog box is now shown to allow users to choose where to save the MCA data.
MCA data is now also written into a text file in column format (i.e. values
for each element are in separate columns) to allow for easy plotting using
external tools. The file name is the same as the .mca file, with '.dat'
appended to it.

Change-Id: I05b66c32971a215f526d08ad4d49c1f2e7c901aa