Replace printStackTrace with logging
printStackTrace dumps the stack trace to the stdout, on a beamline
running with the standard startup bash scripts this is captured into a
log file. As this avoids the proper logging system these files are not
rolled over or compressed so they can get large which doesn't please our
sys admins.

Change-Id: I147cba20609aecc80db30137581e05cd1c602453