DAQ-470 Load LocalProperties and pass to Log/Channel Servers
Load local properties explicitly up front in initialiseServerEnvironment
in Configuration Defaults. This is one of the first methods to run on
startup via the start method of GDAServerApplication. Doing so allows
the properties required by the Log/Channel and Name servers to be passed
to them rather than each loading all the properties from the filesystem.

The server command builder methods are updated to accommodate this. Also
add some convenience methods to translate the CORBA key names before
giving them to the ChannelServer and help with array/varargs mangling.

Add logging too and tidy up some of the enum defs.

Change-Id: Id70eaad7829c18b593a9b18d8aef3c8aae381b5b